@@ -11,14 +11,19 @@
     "start": "webpack-dev-server --open"
   },
   "dependencies": {
+    "css-loader": "^5.0.1",
     "history": "^4.10.1",
+    "node-sass": "^5.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.3",
     "react-router-dom": "^5.1.2",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "sass": "^1.32.2",
+    "sass-loader": "^10.1.0",
+    "style-loader": "^2.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.4.3",
@@ -27,7 +32,7 @@
     "babel-loader": "^8.0.5",
     "html-webpack-plugin": "^3.2.0",
     "path": "^0.12.7",
-    "webpack": "^4.29.6",
+    "webpack": "^4.45.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "^3.2.1"
   }
