@@ -1,56 +1,39 @@
-import React from 'react';
-import { Router, Route, Switch, Redirect } from 'react-router-dom';
-import { connect } from 'react-redux';
+import React from "react";
+import { Router, Route, Switch, Redirect } from "react-router-dom";
+import { useSelector, useDispatch } from "react-redux";
 
-import { history } from '../_helpers';
-import { alertActions } from '../_actions';
-import { PrivateRoute } from '../_components';
-import { HomePage } from '../HomePage';
-import { LoginPage } from '../LoginPage';
-import { RegisterPage } from '../RegisterPage';
+import { history } from "../_helpers";
+import { alertActions } from "../_actions";
+import { PrivateRoute } from "../_components";
+import { Application } from "../Application";
+import { LoginPage } from "../LoginPage";
+import { RegisterPage } from "../RegisterPage";
 
-class App extends React.Component {
-    constructor(props) {
-        super(props);
+import './App.scss';
 
-        history.listen((location, action) => {
-            // clear alert on location change
-            this.props.clearAlerts();
-        });
-    }
 
-    render() {
-        const { alert } = this.props;
-        return (
-            <div className="jumbotron">
-                <div className="container">
-                    <div className="col-sm-8 col-sm-offset-2">
-                        {alert.message &&
-                            <div className={`alert ${alert.type}`}>{alert.message}</div>
-                        }
-                        <Router history={history}>
-                            <Switch>
-                                <PrivateRoute exact path="/" component={HomePage} />
-                                <Route path="/login" component={LoginPage} />
-                                <Route path="/register" component={RegisterPage} />
-                                <Redirect from="*" to="/" />
-                            </Switch>
-                        </Router>
-                    </div>
-                </div>
-            </div>
-        );
-    }
-}
+export const App = () => {
+  const dispatch = useDispatch();
+  const alert = useSelector((state) => state.alert);
 
-function mapState(state) {
-    const { alert } = state;
-    return { alert };
-}
+  const clearAlerts = alertActions.clear;
 
-const actionCreators = {
-    clearAlerts: alertActions.clear
+  history.listen((location, action) => {
+    dispatch(clearAlerts());
+  });
+  return (
+    <div>
+      {alert.message && (
+        <div className={`alert ${alert.type}`}>{alert.message}</div>
+      )}
+      <Router history={history}>
+        <Switch>
+          <PrivateRoute exact path="/trade" component={Application} />
+          <Route path="/login" component={LoginPage} />
+          <Route path="/register" component={RegisterPage} />
+          <Redirect from="*" to="/trade" />
+        </Switch>
+      </Router>
+    </div>
+  );
 };
-
-const connectedApp = connect(mapState, actionCreators)(App);
-export { connectedApp as App };
