@@ -3,7 +3,7 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
 import { store } from './_helpers';
-import { App } from './App';
+import {App} from './App';
 
 render(
     <Provider store={store}>
